Refactored previewWindow components
- Merged Invoice component into mainContent component
- Move all event handlers from mainContent to Viewer
- Viewer now connect to translate HOC and pass props down to sideBar and mainContent
